+++ masterSolution/PolyWars/Model/Triangle.cs
@@ -30,7 +30,7 @@ namespace PolyWars.Model{
             return new PointCollection() { point1, point2, point3 };
         }
 
-        public int RPM { get; set; }
+        public double RPM { get; set; }
         public int VerticalSpeed { get; set; }
         public int HorizontialSpeed { get; set; }
